Add Task class
Task contains a String description and a boolean isDone.
It has a constructor that takes in a String description parameter, as well as setting isDone to false by default.
getStatusIcon() creates a visual for its "done" status.
setAsDone() and setAsUndone are setter functions that sets the isDone variable to true and false respectively.
Overrode the default Object.toString() so that the Task instances displays completion status and task description / name.
Changed the ArrayList type in RedGirl from String to Task.